a demo of adding two simple arrays
in theory, i would be able to add arrays of a size up to 512, however i do not have enough pins to test that, not sure if there is a better way that would let me do that. Also, sending the results AND reading them happens sequentially. It is not very difficult to make them write the arrays and read the results in parallel but I did not bother since I cannot make large enough arrays to test that anyways, and it would add a lot of complexity. however, the part where it's doing the arithmetic itself, the ppu_controller, itself does operation in parallel as it has two ALUs. But it also doesn't do this unless there are large enough inputs